requirements.txt
# Core ML and Computer Vision
torch>=1.12.0
torchvision>=0.13.0
ultralytics>=8.0.0
opencv-python>=4.5.0
numpy>=1.21.0
scipy>=1.7.0
Pillow>=8.3.0

# 3D Processing and Rendering
trimesh>=3.9.0
matplotlib>=3.5.0

# Data Processing
pandas>=1.3.0
scikit-learn>=1.0.0

# Configuration and Utilities
pyyaml>=6.0
tqdm>=4.62.0
click>=8.0.0
loguru>=0.6.0

# Model-specific dependencies
diffusers>=0.21.0  # For Marigold depth estimation
transformers>=4.30.0  # For Marigold and Detic
accelerate>=0.20.0  # For Marigold
einops>=0.6.0  # For modern ML models
timm>=0.9.0  # For vision models
xformers>=0.0.20  # For efficient attention (optional)

# Detectron2 and related
detectron2>=0.6  # For Mask R-CNN and Detic
fvcore>=0.1.5  # Detectron2 dependency
iopath>=0.1.9  # Detectron2 dependency

# Additional ML libraries
tensorboard>=2.10.0  # For training visualization
wandb>=0.13.0  # For experiment tracking (optional)
